backend/requirements.txt
numpy
pandas
pyarrow
pydantic
pydantic_core
fastapi
pydantic
uvicorn
fastparquet
python-multipart
psycopg2-binary
asyncpg
databases
redis
pydantic_settings
requests
clickhouse_sqlalchemy
greenlet
pyjwt
passlib
